Refactor errors to _errors.py
