+++ src/python/grpcio/grpc/BUILD.bazel
@@ -100,6 +100,11 @@ py_library(
 )
 
 py_library(
+    name = "errors",
+    srcs = ["_errors.py"],
+)
+
+py_library(
     name = "grpcio",
     srcs = ["__init__.py"],
     data = [
@@ -115,6 +120,7 @@ py_library(
         ":auth",
         ":channel",
         ":compression",
+        ":errors",
         ":interceptor",
         ":plugin_wrapping",
         ":server",
