+++ src/python/grpcio/grpc/_cython/_cygrpc/aio/callback_common.pyx.pxi
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from grpc._errors import InternalError
 
 cdef class CallbackFailureHandler:
