+++ src/python/grpcio/grpc/_cython/_cygrpc/aio/channel.pyx.pxi
@@ -13,6 +13,7 @@
 # limitations under the License.
 #
 
+from grpc._errors import UsageError
 
 class _WatchConnectivityFailed(Exception):
     """Dedicated exception class for watch connectivity failed.
