+++ src/python/grpcio/grpc/_cython/_cygrpc/aio/server.pyx.pxi
@@ -13,6 +13,8 @@
 # limitations under the License.
 
 
+from grpc._errors import BaseError, AbortError, InternalError, UsageError
+
 import inspect
 import traceback
 import functools
