+++ src/python/grpcio/grpc/_cython/cygrpc.pyx
@@ -55,7 +55,6 @@ include "_cygrpc/tag.pyx.pxi"
 include "_cygrpc/time.pyx.pxi"
 include "_cygrpc/vtable.pyx.pxi"
 include "_cygrpc/_hooks.pyx.pxi"
-include "_cygrpc/common.pyx.pxi"
 include "_cygrpc/observability.pyx.pxi"
 
 include "_cygrpc/grpc_gevent.pyx.pxi"
