+++ src/python/grpcio/grpc/_errors.py
@@ -1,4 +1,4 @@
-# Copyright 2023 The gRPC Authors
+# Copyright 2024 The gRPC authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,29 +12,46 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+#############################  gRPC Exceptions  ################################
+
+
 class BaseError(Exception):
     """
     The base class for exceptions generated by gRPC.
     """
 
+
 class UsageError(BaseError):
     """
     Raised when the usage of API by applications is inappropriate.
-
     For example, trying to invoke RPC on a closed channel, mixing two styles
     of streaming API on the client side. This exception should not be
     suppressed.
     """
 
+
 class AbortError(BaseError):
     """
     Raised when calling abort in servicer methods.
-
     This exception should not be suppressed. Applications may catch it to
     perform certain clean-up logic, and then re-raise it.
     """
+
 
 class InternalError(BaseError):
     """
     Raised upon unexpected errors in native code.
     """
+
+
+class RpcError(BaseError):
+    """Raised by the gRPC library to indicate non-OK-status RPC termination."""
+
+
+__all__ = (
+    "BaseError",
+    "UsageError",
+    "AbortError",
+    "InternalError",
+    "RpcError",
+)
